Created separated modules for each object
