+++ vip_hci/__init__.py
@@ -10,6 +10,6 @@ from . import fm
 from . import metrics
 from . import stats
 from . import var
+from .objects import *
 from .dataset import *
-from .postproc import *
 from .vip_ds9 import *
